Add files via upload
This is implemented using VB.NET and SQL server. This mainly consists of speech synthesis which reads the data to the user , the data can be an XML file or text file. Later the text can be saved as .wav format. Next is a math Game app which is voice oriented where random numbers and random operator is generated to perform basic of mathematics. It consists of levels where level 2 is a timer based. This makes the user feels bilaterality where the voice and sounds are played at the background. Next is a sticky note , a Sticky note with a speech facility, the text is read to the user further the file can be uploaded  into the database.

This application uses the TEXT TO SPEECH [TTS] feature in all the modules even the login and registration forms are speech oriented. This aims to meet all kinds of people even with visually impaired, native language speaker and older people to make their interaction with the application interactively.